Firebase / fireStore Teste gerade denn geter

+++ app/src/main/java/com/example/abschlussaufgabe/ui/MaterialReceivedFragment.kt
@@ -91,7 +91,7 @@ class MaterialReceivedFragment : Fragment() {
                 viewModel.checkMaterialId(id)
 
                 //udate StorageMaterial Model Datenbank
-                viewModel.updateMaterialLocation(id, viewModel.userData.userId)
+                viewModel.updateMaterialLocation(id, viewModel.userData.userId.toInt())
 
                 //Spile sound ab wenn  material empfange
                 viewModel.playActionSound(context!!)
